minor improve

+++ src/main/java/cn/ussshenzhou/madparticle/particle/optimize/InstancedRenderManager.java
@@ -104,7 +104,7 @@ public class InstancedRenderManager {
     private static void resetOitTexture() {
         var window = Minecraft.getInstance().getWindow();
         glBindTexture(GL_TEXTURE_2D, ACCUM_TEXTURE);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, window.getWidth(), window.getHeight(), 0, GL_RGBA, GL_HALF_FLOAT, (ByteBuffer) null);
+        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, window.getWidth(), window.getHeight(), 0, GL_RGBA, GL_HALF_FLOAT, (ByteBuffer) null);
         glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
         glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
         glBindTexture(GL_TEXTURE_2D, REVEAL_TEXTURE);
